requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
python-multipart>=0.0.9,<1.0
numpy>=1.24,<3.0
opencv-python>=4.9,<5.0
mediapipe>=0.10.14,<0.11
joblib>=1.3,<2.0
scikit-learn>=1.4,<2.0
pillow>=10.0,<12.0
pymongo>=4.8,<5.0
tensorflow>=2.14,<2.19; python_version < "3.13"
torch>=2.2,<3.0
torchvision>=0.17,<1.0
timm>=1.0,<2.0
playsound==1.2.2
scipy>=1.10,<2.0
pytest>=8.3,<9.0
httpx>=0.27,<1.0
